requirements.txt
yt-dlp
mutagen
Pillow
PyYAML
requests
